pyproject.toml
[tool.poetry]
name = "empty"
version = "0.1.0"
description = ""
authors = ["R.Mammeri <user.medmammeri@gmail.com>"]
readme = "README.md"
packages = [
    { include = "empty" }
    ]
[tool.poetry.dependencies] # main dependency group
python = "^3.8.1"
mkdocs-blog-plugin = "^0.25"


[tool.poetry.group.docs.dependencies]
mkdocs-material = "^9.2.0b0"
mkdocs-enumerate-headings-plugin = "^0.6.1"
mkdocstrings = {extras = ["python"], version = "^0.22.0"}
mkdocs-gen-files = "^0.5.0"
mkdocs-literate-nav = "^0.6.0"
minify = "^66.0.2"
mkdocs-minify-plugin = "^0.6.4"
mkdocs-git-revision-date-localized-plugin = "^1.2.0"
latex2markdown = "^0.2.1"

[tool.poetry.group.dev.dependencies]
pydantic = "^1.10.8"
python-dotenv = "^1.0.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.1.0"

[tool.poetry.group.lint.dependencies]
mypy = "^1.3.0"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
colorama = "^0.4.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# pytest options
[tool.pytest.ini_options]
addopts = '''
--cov=my_package
--cov-report=term-missing
'''
testpaths = "tests"

# coverge reporting options
[tool.coverge.report]
exclude_line = [
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING",
    "if typing.TYPE_CHECKING",
    "if t.TYPE_CHECKING",
    "@overload",
    "@typing.overload",
    "@t.overload",
    "pragma: no cover",
]

[tool.coverage.run]
omit = [
    "*/tests/*",
]

[tool.black]
line-length=100
color = true
exclude = '''
/(
    \.git
    | \.hg
    | notebooks
    | \.mypy_cache
    | \.tox
    | \.nox
    | \.venv
    | _build
    | \.eggs
    | buck-out
    | build
    | dist
    | env
    | venv
    | \.svn
)/
'''

[tool.isort]
# to match black style
line_length = 100

known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
multi_line_output = 3
include_trailing_comma = true
profile = "black"
force_grid_wrap = 0
use_parentheses = true
indent = 4
ensure_newline_before_comments = true
color_output = true

# google style guide for imports
force_single_line = true
force_sort_within_sections = true
lexicographical = true
order_by_type = true
single_line_exclusions = [
    "typing",
]

known_first_party = [
    "my_package",
    "tests",
]

[tool.mypy]
exclude = [
    "venv/"
]
pretty = true
show_traceback = true
color_output = true

allow_redefinition = false
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
ignore_missing_imports = true
implicit_reexport = false
no_implicit_optional = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true